+++ app/tag/controller.js
@@ -1,7 +1,15 @@
 const Tag = require("./model");
+const { policyFor } = require("../policy");
 
 const store = async (req, res, next) => {
   try {
+    let policy = policyFor(req.user);
+    if (!policy.can("create", "Tag")) {
+      return res.json({
+        error: 1,
+        message: "Anda tidak memiliki akses untuk membuat tag",
+      });
+    }
     let payload = req.body;
 
     let tag = new Tag(payload);
@@ -22,6 +30,12 @@ const store = async (req, res, next) => {
 
 const update = async (req, res, next) => {
   try {
+    if (!policy.can("update", "Tag")) {
+      return res.json({
+        error: 1,
+        message: "Anda tidak memiliki akses untuk mengupdate tag",
+      });
+    }
     let payload = req.body;
 
     let tag = await Tag.findOneAndUpdate({ _id: req.params.id }, payload, {
@@ -44,6 +58,12 @@ const update = async (req, res, next) => {
 
 const destroy = async (req, res, next) => {
   try {
+    if (!policy.can("delete", "Tag")) {
+      return res.json({
+        error: 1,
+        message: "Anda tidak memiliki akses untuk menghapus tag",
+      });
+    }
     let deleted = await Tag.findOneAndDelete({ _id: req.params.id });
     return res.json(deleted);
   } catch (error) {
